Customer meta hardening
Added documentation describing customer profile implentation to customer.php

Added function to allow last active time to be set for users other than the current user.

Avoid deleting users that have posts, comments or purchases associated with thier user id.

Fill in purchase log user id field with temporary profile id when purchase is completed.

Adjusted the temporary profile delete processing and logic to eliminate unused temporary customer
profiles sooner.

Added filters paralleling what is available in WordPress for hooking customer meta CRUD operations.
It would be great to have these filters for ewach of our meta types.

+++ wpsc-includes/cron.php
@@ -3,17 +3,12 @@ add_action( 'wpsc_hourly_cron_task', 'wpsc_clear_stock_claims' );
 add_action( 'wpsc_hourly_cron_task', '_wpsc_clear_customer_meta' );
 
 /**
- * Clears the stock claims, runs on hourly WP_Cron event and when editing purchase log statuses.
- *
- * @since 3.8.9
- * @access public
- *
- * @return void
+ * wpsc_clear_stock_claims, clears the stock claims, runs using wp-cron and when editing purchase log statuses via the dashboard
  */
 function wpsc_clear_stock_claims() {
 	global $wpdb;
 
-	$time     = (float) get_option( 'wpsc_stock_keeping_time', 1 );
+	$time = (float) get_option( 'wpsc_stock_keeping_time', 1 );
 	$interval = get_option( 'wpsc_stock_keeping_interval', 'day' );
 
 	// we need to convert into seconds because we're allowing decimal intervals like 1.5 days
@@ -23,43 +18,52 @@ function wpsc_clear_stock_claims() {
 		'week' => 604800,
 	);
 
-	$seconds = floor( $time * $convert[ $interval ] );
+	$seconds = floor( $time * $convert[$interval] );
 
 	$sql = $wpdb->prepare( "DELETE FROM " . WPSC_TABLE_CLAIMED_STOCK . " WHERE last_activity < UTC_TIMESTAMP() - INTERVAL %d SECOND", $seconds );
 	$wpdb->query( $sql );
 }
 
-/**
- * Purges customer meta that is older than WPSC_CUSTOMER_DATA_EXPIRATION on an hourly WP_Cron event.
- *
- * @since 3.8.9.2
- * @access public
- *
- * @return void
- */
 function _wpsc_clear_customer_meta() {
 	global $wpdb;
 
 	require_once( ABSPATH . 'wp-admin/includes/user.php' );
 
-	$purge_count = 200;
 
-	$sql = "
-		SELECT user_id
-		FROM {$wpdb->usermeta}
-		WHERE
-		meta_key = '_wpsc_last_active'
-		AND meta_value < UNIX_TIMESTAMP() - " . WPSC_CUSTOMER_DATA_EXPIRATION . "
-		LIMIT {$purge_count}
-	";
+	$args = array(
+			'meta_query' => array(
+					array(
+							'key'     => _wpsc_get_customer_meta_key( 'temporary_profile' ),
+							'value'   => time(),
+							'type'    => 'numeric',
+							'compare' => '<'
+					)
+			),
+			'fields' => 'ID'
+	);
 
-	/* Do this in batches of 200 to avoid memory issues when there are too many anonymous users */
-	@set_time_limit( 0 ); // no time limit
+	 $wp_user_query = new WP_User_Query( $args );
 
-	do {
-		$ids = $wpdb->get_col( $sql );
-		foreach ( $ids as $id ) {
+
+	// For each of the ids double check to be sure there isn't any important data associated with the temporary user.
+	// If important data is found the user is no longer temporary. We also use a filter so that if other plug-ins
+	// want to either stop the user from being deleted, or do something with the information in the profile they
+	// have that chance.
+	foreach ( $wp_user_query->results as $id ) {
+		// for extra safety
+		$ok_to_delete_temporary_customer_profile = ( wpsc_customer_purchase_count( $id ) == 0 ) && ( wpsc_customer_post_count( $id ) == 0 ) && ( wpsc_customer_comment_count( $id ) == 0 );
+		if ( apply_filters( 'wpsc_before_delete_customer_profile', $ok_to_delete_temporary_customer_profile, $id ) ) {
 			wp_delete_user( $id );
+			do_action( 'wpsc_after_delete_temp_customer_profile', $id );
+		} else {
+			// user should not be temporary if it has posts, purchases, comments or anything else.  This is partially a
+			// defensive measure against the list of temporary users growing forever should there be logic problems
+			// with other plug-ins and their implementation of the wpsc_before_delete_customer_profile filter.
+			wpsc_delete_customer_meta( 'temporary_profile' );
+			do_action( 'wpsc_customer_profile_not_temporary', $id );
 		}
-	} while ( count( $ids ) == $purge_count );
+	}
 }
+
+add_action( 'testclearmeta' , _wpsc_clear_customer_meta );
+
